+++ src/StateSmithTest/spec/2/c/Spec2Sm.c
@@ -11251,7 +11251,7 @@ static void TEST9B_ROOT__A4_ev1(Spec2Sm* sm)
         // uml: / { trace("Transition action `` for TEST9B_ROOT__A3.ExitPoint(1) to TEST9B_ROOT__B4."); } TransitionTo(TEST9B_ROOT__B4)
         {
             // Step 1: Exit states until we reach `TEST9B_ROOT` state (Least Common Ancestor for transition).
-            TEST9B_ROOT__A3_exit(sm);
+            exit_up_to_state_handler(sm, TEST9B_ROOT_exit);
             
             // Step 2: Transition action: `trace("Transition action `` for TEST9B_ROOT__A3.ExitPoint(1) to TEST9B_ROOT__B4.");`.
             trace("Transition action `` for TEST9B_ROOT__A3.ExitPoint(1) to TEST9B_ROOT__B4.");