+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- fix specification bug - exit point transition only exits parent (not ancestors).
+  - https://github.com/StateSmith/StateSmith/issues/132
+
+---
+
 ## [0.8.5-alpha]
 ### Added
 - `TimeStringParser` - very handy to use in expansions so that you can write guard clauses like `[ t1After(3 min) ]`.